backend/requirements.txt
Psycopg2-binary
Flask
sockets
